Add version checking (#171)
This is a partial copy of the outdated PR of @HarmvZ who laid the
foundations. However, it now also checks the latest supported version.

The original PR had some problems with double running the check and the
new check needed access to the `base_url` anyway. So I decided put the
checking in the `ClientBase` initialisation but keep it in a separate
httpx Client.

---------

Co-authored-by: HarmvZ <harm.van.zeeland@gmail.com>
Co-authored-by: James Meakin <12661555+jmsmkn@users.noreply.github.com>