Add Create Strategies (#173)
Part of the large PR:
- https://github.com/DIAGNijmegen/rse-gcapi/pull/167

This PR adds creation strategies for creating interdependent objects
that can be used for creating display sets and archive items. Strategies
_should_ return an item that can be used to finalize the strategy, the
CIV strategies return a `ComponentInterfaceValuePostRequest`.

This PR implements the logic behind the following:

## Allowing both `str`, and `list[str]`

Sourcing a CIV from one file no longer requires a one-lengthed list, the
helper functions automagically handle both. This also applies to
pathlib.Path.

## RE-use of images / CIVs

Can provide a CIV or Image to the strategies and it will be re-used.