SocketValueSpec: update and fix integration tests (#242)
The integration tests now either only read or write
without checking for consequences for the latter. This removes a large
part of the testing utilities since we no longer probe the integration
GC instance if things have imported. As long as the response from GC is
OK we consider the create/update test to pass.

In addition, the integration tests have been expanded to also check the
reusing of images and socket values as a whole for each of the main
super kinds.